test: move _scout_list_response() to conftest.py and reuse it in test_server.py
test_formatters.py already had a private _scout_list_response() builder for
the list_scouts response shape. test_server.py independently hand-wrote the
identical {"scouts": [], "total": 0, "summary": {...}} literal in two tests
that mock the adapter's return value for the same call. Moving the builder
into conftest.py (previously just a placeholder) lets both files share one
source of truth instead of drifting if the shape ever changes.

Test-only, no behavior change: same dicts, same assertions. Verified via
uv run pytest -q (445 passed/1 skipped, matching baseline) and
uv run ruff check . clean.
